+++ src/cjs/CircleSector.d.ts
@@ -8,13 +8,14 @@
  * @modified 2024-03-08 Added the `containsAngle` method.
  * @modified 2024-03-09 Added the `circleSectorIntersection` method to find coherent sector intersections..
  * @modified 2024-03-09 Added the `angleAt` method to determine any angle at some ratio.
+ * @modified 2025-04-02 Adding `CircleSector.lineIntersections` and `CircleSector.lineIntersectionTangents` and implementing `Intersectable`.
  * @version  1.2.0
  **/
 import { Circle } from "./Circle";
 import { Vector } from "./Vector";
 import { VertTuple } from "./VertTuple";
 import { Vertex } from "./Vertex";
-import { SVGPathParams, SVGSerializable, UID, XYCoords } from "./interfaces";
+import { Intersectable, SVGPathParams, SVGSerializable, UID, XYCoords } from "./interfaces";
 /**
  * @classdesc A simple circle sector: circle, start- and end-angle.
  *
@@ -24,7 +25,7 @@ import { SVGPathParams, SVGSerializable, UID, XYCoords } from "./interfaces";
  * @requires UIDGenerator
  * @requires XYCoords
  **/
-export declare class CircleSector implements SVGSerializable {
+export declare class CircleSector implements Intersectable, SVGSerializable {
     /**
      * Required to generate proper CSS classes and other class related IDs.
      **/