+++ src/cjs/utils/helpers/VEllipseHelper.d.ts
@@ -1,10 +1,10 @@
 /**
  * A helper for VEllipses.
  *
- * @author  Ikaros Kappler
- * @date    2025-03-31
+ * @author   Ikaros Kappler
+ * @date     2025-03-31
  * @modified 2025-04-02 Adding `VEllipseHelper.drawHandleLines()`.
- * @version 1.0.0
+ * @version  1.0.0
  */
 import { VEllipse } from "../../VEllipse";
 import { Vertex } from "../../Vertex";