+++ src/ts/utils/helpers/VEllipseSectorHelper.ts
@@ -1,23 +1,35 @@
 /**
  * A helper for VEllipseSectors.
  *
- * @author  Ikaros Kappler
- * @date    2021-03-24
- * @version 1.0.0
+ * @author   Ikaros Kappler
+ * @date     2021-03-24
+ * @modified 2025-04-02 Adding `VEllipseSectorHelper.drawHandleLines`.
+ * @version  1.1.0
  */
 
 import { Line } from "../../Line";
 import { VEllipseSector } from "../../VEllipseSector";
 import { Vertex } from "../../Vertex";
 import { VertEvent } from "../../VertexListeners";
+import { DrawLib } from "../../interfaces";
 
 export class VEllipseSectorHelper {
+  private sector: VEllipseSector;
+  private startAngleControlPoint: Vertex;
+  private endAngleControlPoint: Vertex;
+  private rotationControlPoint: Vertex;
+
   constructor(
     sector: VEllipseSector,
     startAngleControlPoint: Vertex,
     endAngleControlPoint: Vertex,
     rotationControlPoint: Vertex
   ) {
+    this.sector = sector;
+    this.startAngleControlPoint = startAngleControlPoint;
+    this.endAngleControlPoint = endAngleControlPoint;
+    this.rotationControlPoint = rotationControlPoint;
+
     const rotationControlLine: Line = new Line(sector.ellipse.center, rotationControlPoint);
     const startAngleControlLine: Line = new Line(sector.ellipse.center, startAngleControlPoint);
     const endAngleControlLine: Line = new Line(sector.ellipse.center, endAngleControlPoint);
@@ -35,8 +47,8 @@ export class VEllipseSectorHelper {
     // | Listen for rotation changes.
     // +-------------------------------------------
     rotationControlPoint.listeners.addDragListener((event: VertEvent) => {
-      var newRotation = rotationControlLine.angle();
-      var rDiff = newRotation - sector.ellipse.rotation;
+      const newRotation: number = rotationControlLine.angle();
+      const rDiff: number = newRotation - sector.ellipse.rotation;
       sector.ellipse.rotation = newRotation;
       sector.ellipse.axis.rotate(rDiff, sector.ellipse.center);
       startAngleControlPoint.rotate(rDiff, sector.ellipse.center);
@@ -55,6 +67,28 @@ export class VEllipseSectorHelper {
     // +-------------------------------------------
     endAngleControlPoint.listeners.addDragListener((event: VertEvent) => {
       sector.endAngle = endAngleControlLine.angle() - sector.ellipse.rotation;
+    });
+  }
+
+  /**
+   * Draw grey handle lines.
+   *
+   * @param {DrawLib<any>} draw - The draw library instance to use.
+   * @param {DrawLib<any>} fill - The fill library instance to use.
+   */
+  drawHandleLines(draw: DrawLib<any>, fill: DrawLib<any>) {
+    draw.line(this.sector.ellipse.center, this.startAngleControlPoint, "rgba(64,192,128,0.333)", 1.0, {
+      dashOffset: 0.0,
+      dashArray: [4, 2]
+    });
+    draw.line(this.sector.ellipse.center, this.endAngleControlPoint, "rgba(64,192,128,0.333)", 1.0, {
+      dashOffset: 0.0,
+      dashArray: [4, 2]
+    });
+    // Draw rotation handle line
+    draw.line(this.sector.ellipse.center, this.rotationControlPoint, "rgba(64,192,128,0.333)", 1.0, {
+      dashOffset: 0.0,
+      dashArray: [4, 2]
     });
   }
 }